Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "AI backend for ovarian cyst prediction and treatment"
authors = ["Achieng Mary, Laura Gachanja"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = "^0.22.0"
requests = "^2.31.0"
pydantic = "^2.0"
joblib = "^1.5.1"
pandas = "^2.3.1"
scikit-learn = "^1.7.0"
xgboost = "^3.0.2"
pinecone = "^7.3.0"
python-dotenv = "^1.1.1"
numpy = "^1.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
